requirements.txt
pytorch==1.8.1
torchvision==0.9.1
pytest==6.2.4
